Feat: Simple OpenGraph tags (#184)
