+++ src/app/proposals/[id]/_components/participants.component.tsx
