+++ src/app/providers.tsx
@@ -1,3 +1,4 @@
+"use client";
 import { ReactNode, useEffect, useState } from "react";
 import { WagmiProvider } from "wagmi";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
